loads of updates

+++ src/nvim-plugins/kitty-open-file/lua/kitty_open_file.lua
@@ -2,8 +2,11 @@ return {
   setup = function()
     local dir = vim.fn.stdpath("state") .. "/kitty-remotes"
     vim.fn.system("mkdir -p " .. dir)
-    local filename = dir .. "/window-" .. os.getenv("KITTY_WINDOW_ID")
-    vim.fn.system("rm -f " .. filename)
-    vim.fn.serverstart(filename)
+
+    if os.getenv("KITTY_WINDOW_ID") then
+      local filename = dir .. "/window-" .. os.getenv("KITTY_WINDOW_ID")
+      vim.fn.system("rm -f " .. filename)
+      vim.fn.serverstart(filename)
+    end
   end,
 }
